refactor: changing double to long double

+++ src/libs/libs/graph.h
@@ -10,9 +10,9 @@ void graph_destruct(Graph *graph);
 
 int graph_get_num_nodes(Graph *graph);
 
-double graph_get_cost_from_edge(Graph *graph, int src, int dest);
+long double graph_get_cost_from_edge(Graph *graph, int src, int dest);
 
-void graph_add_edge(Graph *graph, int src, int dest, double cost);
+void graph_add_edge(Graph *graph, int src, int dest, long double cost);
 
 void graph_read(Graph *graph, FILE *file, int num_edges);
